pyproject.toml
[tool.poetry]
name = "pyobs-zaber"
version = "1.0.3"
description = "pyobs model for Zaber motors"
authors = ["Leon Meerwart <leon.meerwart@uni-goettingen.de>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
zaber-motion = "^4.7.0"
pyobs-core = "^1.6.7"

[tool.poetry.dev-dependencies]
black = ">21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']